+++ CHANGELOG.md
@@ -5,6 +5,8 @@
 ### New features and improvements
 
 - [CDTS](https://gccode.ssc-spc.gc.ca/iitb-dgiit/nw-ws/sgdc-cdts) v5.0.4 & [wet-boew](https://github.com/wet-boew/wet-boew) v4.0.85
+- Adding an optional signInUrl property for Session Timeout allowing users to specify the sign-in page url when the session has expired
+- Adding an optional textOverrides property for Session Timeout allowing users to override specific text elements
 - Updating static fallback files
 
 ## v2.12.0